Add endpoint override
